Add MessageContainerView to restructure MessageItemView
